Simplify llama-server startup process by removing detached mode (tested on macOS and Windows, parent process termination does not affect child process)
